Checklist (#35)
* downloaded necessary installations and made a new api key, added env file and configured to a past event

* started checklist section

was able to add chekcboxes and drop down boxes but it is in the upper right corner of the page, not in the proper orientation or font... need to edit style page to accomoadte the new list

* Continued adding to checklist

Added more checks and continued through the document. Got to Full Systems Check in the enabled section. Will continue to type these and then proceed to styling in the styles.css file.

* Changed source of buttons
Instead of hardcoding the checklist, I am now inserting it into a JSON file to make it easier to change. Mr. S and I went over a format to use so I will copy the same ideas in the JSON file and then figure out how to insert it into the .ejs file, starting small with just Post-Match( in pit) and Robot Items.

* Redid checklist to ensure easy edits

I did the checklist area through the JSON file and changed the ejs file to adapt to what the JSON file says. I have added two sections and I have made it styled so it is where it is supposed to be based on the past groups styles.

* Got JSON working and added in a new property called checked

Mapped each object in the JSON file to an object and to have the property checked within teh object. Need to work on robot.js to make the query selector work

* Made sure the checkboxes worked

It now prints out its name to the console when clicked, confirming the function of the actionListener.

* started web scokets
i made it so that there is a use of web sockets and now i need to update that in router.js and verify it works

* Continued working on web sockets

I continued to work on web sockets but I am still confused on what I need to add in to the code to make it work and where.

* still tryign to get the websocket functions to work

i am confused on how it is supposed to communicate with each other but mr. s explained it to me so hopefully, tomorrow goes better

* Worked on debugging

I cannot get updateChecked to work so I am using the console to work on debugging the actual function to get it to update checked on the right task

* Implement localStorage for checklist persistence

Added data-key attributes to checkboxes and implemented localStorage to save and restore checkbox states across page reloads. Updated the robot.ejs template to generate unique data-key identifiers for each checkbox.

* Changed font in the second page

I changed the font and font size and organization of the second page to be more visually appealing for viewers.

---------

Co-authored-by: diyagupta05 <diyargupta05@gmail.com>